requirements.txt
Flask==2.3.2
Flask-WTF==1.1.1
Werkzeug==2.3.6
WTForms==3.0.1
itsdangerous==2.1.2
numpy==1.24.3
matplotlib==3.7.1
Pillow==9.5.0
python-dotenv==1.0.0
requests==2.31.0
jinja2==3.1.2
pymupdf==1.21.1
pytesseract==0.3.10
wand==0.6.11
streamlit==1.27.0
pypdf2==3.0.1 